LayoutTests/js/arrowfunction-syntax-endings-expected.txt

Tests for ES6 arrow function endings

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS afEOL(12) is 13
PASS x=>x+1 did not throw exception.
PASS x=>x+1
 did not throw exception.
PASS f() is 39
eval('var af = x=>x*2')
PASS af(10) is 20
eval('var af1 = x=>x*3, af2=x=>x*4')
PASS af1(10) is 30
PASS af2(10) is 40
eval('var af1 = x=>x*3;')
PASS af3(10) is 30
eval('var af4 = x=>(x*3)')
PASS af4(10) is 30
eval('var af5 = x=> { return x*3; }')
PASS af5(10) is 30
PASS successfullyParsed is true

TEST COMPLETE
